post security added

+++ src/utils/DeleteFunction.js
@@ -3,7 +3,7 @@ import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation"; //!<<this file, not other
 import { addTableDigits } from "./addTableDigits";
 export default function DeleteButton(nameTable, idData, path, isPost) {
-  async function handleSubmit() {
+  async function handleSubmit(formdata) {
     "use server";
     const db = dbConnect();
     console.log(nameTable.idData);
@@ -11,30 +11,68 @@ export default function DeleteButton(nameTable, idData, path, isPost) {
     const postNumber = Number(nameTable.idData);
     const postTableName = nameTable.nameTable + tableId;
 
-    await db.query(
-      `DELETE FROM ${nameTable.nameTable} WHERE id = ${nameTable.idData} RETURNING *`
-    );
-    const currentPosts = await db.query(
-      `SELECT no_of_posts FROM categories WHERE category_name =  '${nameTable.nameTable}' `
-    );
-    const wrangledPostCount = Number(currentPosts.rows[0].no_of_posts) - 1;
-    await db.query(
-      `UPDATE categories SET no_of_posts = ($1) WHERE category_name = '${nameTable.nameTable}'`,
-      [wrangledPostCount]
-    );
+    const secretCode = formdata.get("wordpass");
 
-    if (nameTable.isPost === 1) {
-      await db.query(`DROP ${postTableName}`);
+    const codeCheck = (
+      await db.query(
+        `SELECT security FROM ${nameTable.nameTable} WHERE id = ${nameTable.idData}`
+      )
+    ).rows;
+    console.log(codeCheck[0].security);
+    console.log(secretCode);
+    if (Number(nameTable.isPost) === 1) {
+      if (secretCode === codeCheck[0].security) {
+        await db.query(
+          `DELETE FROM ${nameTable.nameTable} WHERE id = ${nameTable.idData} RETURNING *`
+        );
+        const currentPosts = await db.query(
+          `SELECT no_of_posts FROM categories WHERE category_name =  '${nameTable.nameTable}' `
+        );
+        const wrangledPostCount = Number(currentPosts.rows[0].no_of_posts) - 1;
+        await db.query(
+          `UPDATE categories SET no_of_posts = ($1) WHERE category_name = '${nameTable.nameTable}'`,
+          [wrangledPostCount]
+        );
+        console.log(nameTable.isPost);
+
+        await db.query(`DROP TABLE IF EXISTS ${postTableName}`);
+      }
+    } else {
+      await db.query(
+        `DELETE FROM ${nameTable.nameTable} WHERE id = ${nameTable.idData} RETURNING *`
+      );
+      const currentPosts = await db.query(
+        `SELECT no_of_posts FROM categories WHERE category_name =  '${nameTable.nameTable}' `
+      );
+      const wrangledPostCount = Number(currentPosts.rows[0].no_of_posts) - 1;
+      await db.query(
+        `UPDATE categories SET no_of_posts = ($1) WHERE category_name = '${nameTable.nameTable}'`,
+        [wrangledPostCount]
+      );
     }
-
     revalidatePath(`/${nameTable.path}`);
     redirect(`/${nameTable.path}`);
   }
+
   return (
     <>
       <form action={handleSubmit}>
         {" "}
         <button className="border-solid border-2 border-red-500">Delete</button>
+        <div>
+          <label htmlFor="password" required>
+            PW:{" "}
+          </label>
+
+          <input
+            className="text-slate-900"
+            type="password"
+            id="user-password"
+            name="wordpass"
+            required
+            placeholder="Case sensitive"
+          />
+        </div>
       </form>
     </>
   );
